Move printBytes to meshUtils
