+++ src/meshUtils.h
@@ -1,4 +1,6 @@
 #pragma once
+#include "DebugConfiguration.h"
+#include <stdint.h>
 
 /// C++ v17+ clamp function, limits a given value to a range defined by lo and hi
 template <class T> constexpr const T &clamp(const T &v, const T &lo, const T &hi)
